requirements.txt
streamlit==1.31.1
python-docx==1.0.1
PyPDF2==3.0.1
nbconvert==7.16.0
beautifulsoup4==4.12.2
notebook>=7.0.0
python-pptx==0.6.23
